Added images to instantiate objects.
Added specific images to each type Bullet, Player, and Meteor objects and added an image function.
Now images do not need to be called randomly in the runner, but instead are instantiated with each class.
This is a big improvement on how images are called to objects.

The next step will be to instantiate sounds with each object, but I still need to get recordings from everyone.
So for now, I will be implementing a feature that allows user to pick the player they want.