Feature added for users to choose player.
Images can now be chosen form players on the front end. Logic had to be created in the
image file to allow unit tests to pass, when there is no css to select from. Code does
need to be refactored.

The next step is to have a starting screen, so the user can select a player before the game starts.