Added start screen and updated wording in jumbotron.
Added a startOption function in the index.js file, to allow users to choose a player before the game starts.
Downside of this is that I did have to initialize ambiguous variables 'var runner' and 'var interval' outside of the class
so that those variables could be used again outside of that function. As of now I do not see any way around calling those variables
once in start option.

Next step will be to add a picture of the player that is chosen on page.